Refactor trace store file backend
